refactor: moved cluster to lab
