+++ src/store/ducks/rootReducer.ts
@@ -2,7 +2,7 @@ import { combineReducers } from 'redux';
 import cowsay from './cowsay';
 import user from './user';
 import labPolicy from './lab/policy';
-import cluster from './cluster';
+import cluster from './lab/cluster';
 
 export default combineReducers({
   cluster,
