+++ src/store/ducks/rootSaga.ts
@@ -4,7 +4,7 @@ import load from './cowsay/sagas';
 import userSagas from './user/sagas';
 import { CowsayTypes } from './cowsay/types';
 import labPolicySagas from './lab/policy/sagas';
-import clusterSagas from './cluster/sagas';
+import clusterSagas from './lab/cluster/sagas';
 
 export default function* rootSaga(): Generator {
   return yield all([
